Add dictionaries for API 3 #1780
- #1780 done
- remove old dictionaries collections and usages
- added indexes to dictionaries collections
- separate dictionaries views from other views
- update load-data script with new dictionaries loading
