+++ server/server/api/views/views.py
@@ -1,7 +1,7 @@
 import json
 import os
 import datetime
-from typing import List, Tuple
+from typing import List
 
 import stripe
 from flask import current_app, request
@@ -15,7 +15,6 @@ from common.mail import send_email
 
 from common.queries import (
     CURRENCIES,
-    DICTIONARIES,
     LANGUAGES,
     MENU,
     SUBMENU,
@@ -26,10 +25,7 @@ from common.queries import (
     IMAGES,
     EPIGRAPHS,
     WHY_WE_READ,
-    DICTIONARYFULL,
     GLOSSARY,
-    DICTIONARY_ADJACENT,
-    DICTIONARY_SIMILAR,
     EXPANSION,
     PWA,
     TRANSLATION_COUNT_BY_DIVISION,
@@ -45,8 +41,6 @@ from common.utils import (
     sort_parallels_key,
     sort_parallels_type_key,
 )
-
-from data_loader.textfunctions import asciify_roman as asciify
 
 from aksharamukha import transliterate
 
@@ -454,62 +448,6 @@ class Parallels(Resource):
         return data, 200
 
 
-class LookupDictionaries(Resource):
-    @cache.cached(key_prefix=make_cache_key, timeout=default_cache_timeout)
-    def get(self):
-        """
-        Send parallel information for given sutta.
-        ---
-        parameters:
-           - in: query
-             name: from
-             type: string
-             required: true
-           - in: query
-             name: to
-             type: string
-           - in: query
-             name: fallback
-             type: string
-        responses:
-            200:
-                schema:
-                    id: dictionary
-                    type: object
-                    properties:
-                        from:
-                            type: string
-                        to:
-                            type: string
-                        dictionary:
-                            type: object
-                            items:
-                                type: array
-                                items:
-                                    type: string
-        """
-        to_lang = request.args.get('to', current_app.config.get('DEFAULT_LANGUAGE'))
-        from_lang = request.args.get('from', None)
-
-        fallback = request.args.get('fallback', 'false')
-        main_dict = False if fallback == 'true' else True
-
-        if from_lang is None:
-            return 'from not specified', 422
-
-        db = get_db()
-
-        result = db.aql.execute(
-            DICTIONARIES,
-            bind_vars={'from': from_lang, 'to': to_lang, 'main': main_dict},
-        )
-
-        try:
-            return result.next(), 200
-        except StopIteration:
-            return 'Dictionary not found', 404
-
-
 class Sutta(Resource):
     @cache.cached(key_prefix=make_cache_key, timeout=default_cache_timeout)
     def get(self, uid, author_uid=''):
@@ -748,74 +686,6 @@ class Glossary(Resource):
         db = get_db()
 
         data = db.aql.execute(GLOSSARY)
-
-        return list(data), 200
-
-
-class DictionaryAdjacent(Resource):
-    @cache.cached(key_prefix=make_cache_key, timeout=default_cache_timeout)
-    def get(self, word=None):
-        """
-        Send list of adjacent terms to dictionary search word
-        ---
-        responses:
-            glossary:
-                type: array
-                items:
-                    type: string
-        """
-        db = get_db()
-
-        data = db.aql.execute(DICTIONARY_ADJACENT, bind_vars={'word': word})
-
-        return list(data), 200
-
-
-class DictionarySimilar(Resource):
-    @cache.cached(key_prefix=make_cache_key, timeout=default_cache_timeout)
-    def get(self, word=None):
-        """
-        Send list of similar terms to dictionary search word
-        ---
-        responses:
-            glossary:
-                type: array
-                items:
-                    type: string
-        """
-        db = get_db()
-
-        data = db.aql.execute(
-            DICTIONARY_SIMILAR, bind_vars={'word': word, 'word_ascii': asciify(word)}
-        )
-
-        return list(data), 200
-
-
-class DictionaryFull(Resource):
-    @cache.cached(key_prefix=make_cache_key, timeout=default_cache_timeout)
-    def get(self, word=None):
-        """
-        Sends list of dictionary entries to dictionary view
-        ---
-        responses:
-            dictionary_full:
-                type: array
-                items:
-                    type: object
-                    properties:
-                        dictname:
-                            type: string
-                        text:
-                            type: string
-
-        """
-        if word is not None:
-            word = word.lower()
-
-        db = get_db()
-
-        data = db.aql.execute(DICTIONARYFULL, bind_vars={'word': word})
 
         return list(data), 200
