+++ server/server/data_loader/dictionaries.py
@@ -1,43 +1,13 @@
+from collections import Counter
 from pathlib import Path
 
+import regex
 from arango.database import Database
 from tqdm import tqdm
 
-from .textfunctions import asciify_roman
+from common.utils import chunks
+from .textfunctions import asciify_roman, pali_sort_key
 from .util import json_load
-
-
-def load_dictionaries(db, dictionaries_dir):
-    load_lookups(db, dictionaries_dir)
-
-
-def load_lookups(db, dictionaries_dir):
-    print('Loading lookup dictionaries')
-    dictionaries = (dictionaries_dir / 'lookup').glob('*.json')
-    dictionaries_collection = db['dictionaries']
-    docs = []
-    for dictionary in tqdm(dictionaries):
-        try:
-            from_lang, to_lang = dictionary.stem.split('2')
-            dict_type = 'maindata'
-            try:
-                to_lang, dict_type = to_lang.split('-')
-            except ValueError:
-                pass
-            main = True if dict_type == 'maindata' else False
-            docs.append(
-                {
-                    'from': from_lang,
-                    'to': to_lang,
-                    'dictionary': json_load(dictionary),
-                    'lookup': True,
-                    'main': main,
-                    'type': 'maindata' if main else dict_type,
-                }
-            )
-        except ValueError:
-            print(f'unknown dictionary name format {dictionary.stem}')
-    dictionaries_collection.import_bulk_logged(docs)
 
 
 def load_simple_dictionaries(db: Database, dictionaries_dir: Path):
@@ -62,19 +32,49 @@ def load_simple_dictionaries(db: Database, dictionaries_dir: Path):
 def load_complex_dictionaries(db: Database, dictionaries_dir: Path):
     complex_dictionaries_dir = dictionaries_dir / 'complex'
     docs = []
+    words_seen = Counter()
+    ids_seen = Counter()
     for dictionary in tqdm(complex_dictionaries_dir.glob('**/*.json')):
         from_lang, to_lang_and_name = dictionary.stem.split('2')
         to_lang, dict_name = to_lang_and_name.split('_')
         content = json_load(dictionary)
         for item in content:
+            word = item['word']
+            words_seen[word] += 1
+
+            # create a meaningful id
+            word_ascii = asciify_roman(word)
+            _key = regex.sub(r'[^a-z0-9]+', '_', word_ascii)
+            if _key in ids_seen:
+                ids_seen[_key] += 1
+                _key += str(ids_seen[_key])
+            else:
+                ids_seen[_key] = 1
+            words_seen[word] += 1
+
             docs.append({
+                '_key': _key,
                 'from': from_lang,
                 'to': to_lang,
                 'word': item['word'],
                 'text': item['text'],
-                'word_ascii': asciify_roman(item['word']),
+                'word_ascii': word_ascii,
                 'dictname': dict_name,
             })
 
-    db.collection('dictionaries_complex').truncate()
-    db.collection('dictionaries_complex').import_bulk(docs)
+    seen = set()
+    for doc in docs:
+        if doc['_key'] in seen:
+            print(f'Duplicate: {doc["_key"]}')
+            seen.add(doc['_key'])
+
+    words_sorted = sorted(words_seen, key=pali_sort_key)
+    word_number = {w: i for i, w in enumerate(words_sorted)}
+
+    for doc in docs:
+        doc['num'] = word_number[doc['word']]
+
+    collection = db.collection('dictionaries_complex')
+    collection.truncate()
+    for chunk in chunks(docs, 1000):
+        collection.import_bulk(chunk, on_duplicate='ignore')
